src/modules/mcp_client/pyproject.toml
[project]
name = "mcp_client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.49.0",
    "mcp>=1.4.1",
    "openai>=1.66.3",
    "python-dotenv>=1.0.1",
]
